+++ com.onesignal.unity.android/Runtime/OneSignalAndroid.cs
@@ -282,12 +282,12 @@ namespace OneSignalSDK {
             return await proxy;
         }
 
-        public override Task<bool> enterLiveActivity(string activityId, string token) {
+        public override Task<bool> EnterLiveActivity(string activityId, string token) {
             SDKDebug.Warn("This feature is only available for iOS.");
             return Task.FromResult(false);
         }
 
-        public override Task<bool> exitLiveActivity(string activityId) {
+        public override Task<bool> ExitLiveActivity(string activityId) {
             SDKDebug.Warn("This feature is only available for iOS.");
             return Task.FromResult(false);
         }